resolve trouble with models
